// Educational.Client/Educational.Client/client/src/api/api.ts
export interface CreateCourseDto {
    id?: string;
    name?: string | undefined;
    description?: string | undefined;
    duration?: number;
    price?: number;
    teacherId?: string;
}
// ...
export interface TeacherForCourseDetailVm {
    userId?: string;
    name?: string | undefined;
}